Makefile: Support incremental builds
Each source file is compiled separately and the resulting object file
is stored in the "build" directory, and these are linked at the end of
compilation to produce the executable.  Rerunning make only rebuilds
object files which are older than the corresponding source file or any
header files it includes.  Dependencies on header files are tracked in
.d files in "build".  This enabled parallel compilation via "make -j N",
where N is the number of simultaneous processes to use.